requirements.txt
streamlit
faiss-cpu
sentence-transformers
PyPDF2
python-docx
langchain
langchain-groq
tiktoken
